backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.5
python-multipart==0.0.6
python-docx==1.1.0
psycopg2-binary==2.9.10
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1
